third_party/blink/web_tests/platform/generic/http/tests/websocket/close-in-order-expected.txt

The closing handshake should start when all data is transmitted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS bufferedAmount is 196608
onclose
PASS bufferedAmount is 0
PASS wasClean is true
PASS reason is "received 196608 bytes"
PASS successfullyParsed is true

TEST COMPLETE
